+++ static/app/components/modals/inviteMembersModal/inviteMembersContext.tsx
@@ -1,22 +1,23 @@
 import {createContext} from 'react';
-import {
+
+import type {
   InviteRow,
   InviteStatus,
   NormalizedInvite,
 } from 'sentry/components/modals/inviteMembersModal/types';
 
 export type InviteMembersContextProps = {
-  willInvite: boolean;
+  complete: boolean;
+  inviteStatus: InviteStatus;
   invites: NormalizedInvite[];
+  pendingInvites: InviteRow[];
+  reset: () => void;
+  sendInvites: () => void;
+  sendingInvites: boolean;
   setEmails: (emails: string[], index: number) => void;
   setRole: (role: string, index: number) => void;
   setTeams: (teams: string[], index: number) => void;
-  sendInvites: () => void;
-  reset: () => void;
-  inviteStatus: InviteStatus;
-  pendingInvites: InviteRow[];
-  sendingInvites: boolean;
-  complete: boolean;
+  willInvite: boolean;
   error?: string;
 };
